+++ edap-protobuf/src/main/java/io/edap/protobuf/ext/codec/BigDecimalCodec.java
@@ -46,10 +46,10 @@ public class BigDecimalCodec implements ExtCodec<BigDecimal> {
     public void encode(ProtoBufWriter writer, BigDecimal bigDecimal) throws EncodeException {
         if (writer.getWriteOrder() == ProtoBufWriter.WriteOrder.SEQUENTIAL) {
             if (bigDecimal == BigDecimal.ZERO) {
-                writer.writeInt32(RANGE_BIGDDECIMAL);
+                writer.writeByte((byte)RANGE_BIGDDECIMAL);
                 writer.writeBytes(new byte[]{1,0,0});
             } else {
-                writer.writeInt32(RANGE_BIGDDECIMAL);
+                writer.writeByte((byte)RANGE_BIGDDECIMAL);
                 byte[] bs = bigDecimal.unscaledValue().toByteArray();
                 writer.writeByteArray(bs, 0, bs.length);
                 writer.writeInt32(bigDecimal.scale(), true);
@@ -57,12 +57,12 @@ public class BigDecimalCodec implements ExtCodec<BigDecimal> {
         } else {
             if (bigDecimal == BigDecimal.ZERO) {
                 writer.writeBytes(new byte[]{1,0,0});
-                writer.writeInt32(RANGE_BIGDDECIMAL);
+                writer.writeByte((byte)RANGE_BIGDDECIMAL);
             } else {
                 byte[] bs = bigDecimal.unscaledValue().toByteArray();
                 writer.writeInt32(bigDecimal.scale(), true);
                 writer.writeByteArray(bs, 0, bs.length);
-                writer.writeInt32(RANGE_BIGDDECIMAL);
+                writer.writeByte((byte)RANGE_BIGDDECIMAL);
             }
         }
