+++ edap-protobuf/src/main/java/io/edap/protobuf/ext/codec/DoubleCodec.java
@@ -37,11 +37,11 @@ public class DoubleCodec implements ExtCodec<Double> {
     @Override
     public void encode(ProtoBufWriter writer, Double v) throws EncodeException {
         if (writer.getWriteOrder() == ProtoBufWriter.WriteOrder.SEQUENTIAL) {
-            writer.writeInt32(RANGE_DOUBLE);
+            writer.writeByte((byte)RANGE_DOUBLE);
             writer.writeFixed64(Double.doubleToRawLongBits(v));
         } else {
             writer.writeFixed64(Double.doubleToRawLongBits(v));
-            writer.writeInt32(RANGE_DOUBLE);
+            writer.writeByte((byte)RANGE_DOUBLE);
         }
     }
 }
