+++ edap-protobuf/src/main/java/io/edap/protobuf/ext/codec/LocalDateCodec.java
@@ -42,11 +42,11 @@ public class LocalDateCodec implements ExtCodec<LocalDate> {
     @Override
     public void encode(ProtoBufWriter writer, LocalDate v) throws EncodeException {
         if (writer.getWriteOrder() == ProtoBufWriter.WriteOrder.SEQUENTIAL) {
-            writer.writeInt32(RANGE_LOCALDATE);
+            writer.writeByte((byte)RANGE_LOCALDATE);
             writer.writeInt32((int)START_DAY.until(v, ChronoUnit.DAYS), true);
         } else {
             writer.writeInt32((int)START_DAY.until(v, ChronoUnit.DAYS), true);
-            writer.writeInt32(RANGE_LOCALDATE);
+            writer.writeByte((byte)RANGE_LOCALDATE);
         }
     }
 }
