+++ edap-protobuf/src/main/java/io/edap/protobuf/ext/codec/LocalTimeCodec.java
@@ -44,11 +44,11 @@ public class LocalTimeCodec implements ExtCodec<LocalTime> {
     @Override
     public void encode(ProtoBufWriter writer, LocalTime v) throws EncodeException {
         if (writer.getWriteOrder() == ProtoBufWriter.WriteOrder.SEQUENTIAL) {
-            writer.writeInt32(RANGE_LOCALTIME);
+            writer.writeByte((byte)RANGE_LOCALTIME);
             writer.writeUInt64(v.atDate(START_DAY).toInstant(ZoneOffset.UTC).toEpochMilli());
         } else {
             writer.writeUInt64(v.atDate(START_DAY).toInstant(ZoneOffset.UTC).toEpochMilli());
-            writer.writeInt32(RANGE_LOCALTIME);
+            writer.writeByte((byte)RANGE_LOCALTIME);
         }
     }
